move auth logic to @/lib
