--- src/app/api/signin/errors.ts
@@ -1,2 +1,0 @@
-export const ERR_USER_NOT_FOUND = new Error("user not found")
-export const ERR_PASSWORD_NOT_MATCH = new Error("invalid email and/or password")
